Move and rename timeline types files (#3473)
It confused the llm, which thought i had a circular dep. really llm just
tried to import everything from .
