+++ packages/sdk/src/index.ts
@@ -74,7 +74,6 @@ export * from './sync-agent/spaces/models/channel'
 export * from './sync-agent/spaces/models/space'
 export * from './sync-agent/spaces/spaces'
 export * from './sync-agent/syncAgent'
-export * from './sync-agent/timeline/models/timeline-types'
 export * from './sync-agent/timeline/timeline'
 export * from './sync-agent/user/models/userInbox'
 export * from './sync-agent/user/models/userMemberships'
@@ -99,6 +98,8 @@ export * from './tests/unit/helpers/ConversationBuilder'
 export * from './types'
 export * from './unauthenticatedClient'
 export * from './utils'
+export * from './views/models/timelineEvent'
+export * from './views/models/timelineTypes'
 export * from './views/streams/channelStreams'
 export * from './views/streams/dmStreams'
 export * from './views/streams/gdmStreams'
@@ -106,7 +107,6 @@ export * from './views/streams/spaceMentionsTransform'
 export * from './views/streams/spaceStreams'
 export * from './views/streams/streamMemberIds'
 export * from './views/streams/streamStatus'
-export * from './views/streams/timelineEvents'
 export * from './views/streams/timelines'
 export * from './views/streams/timelinesModel'
 export * from './views/streams/unreadMarkersTransform'
