+++ packages/sdk/src/tags.ts
@@ -12,7 +12,7 @@ import { bin_fromHexString, bin_toHexString, check } from '@towns-protocol/dlog'
 import { TipEventObject } from '@towns-protocol/generated/dev/typings/ITipping'
 import { isDefined } from './check'
 import { bytesToHex } from 'ethereum-cryptography/utils'
-import { RiverTimelineEvent } from './sync-agent/timeline/models/timeline-types'
+import { RiverTimelineEvent } from './views/models/timelineTypes'
 
 /**
  * Unsafe way to create tags. Possibly used in scenarios where we dont have access to the stream view.
