+++ packages/sdk/src/tests/testDriver_testUtils.ts
@@ -3,7 +3,7 @@ import { DLogger, check, dlog } from '@towns-protocol/dlog'
 import { makeTestClient, makeUniqueSpaceStreamId } from './testUtils'
 import { makeUniqueChannelStreamId } from '../id'
 import { SnapshotCaseType } from '@towns-protocol/proto'
-import { RiverTimelineEvent, TimelineEvent } from '../sync-agent/timeline/models/timeline-types'
+import { RiverTimelineEvent, TimelineEvent } from '../views/models/timelineTypes'
 
 class TestDriver {
     readonly client: Client
