+++ packages/sdk/src/tests/unit/helpers/ConversationBuilder.ts
@@ -9,16 +9,14 @@ import {
     MessageType,
     EventStatus,
     RiverTimelineEvent,
-} from '../../../sync-agent/timeline/models/timeline-types'
-import { ETH_ADDRESS } from '@towns-protocol/web3'
-import { BlockchainTransaction_Tip, PlainMessage } from '@towns-protocol/proto'
-import {
     getIsMentioned,
     getReactionParentId,
     getThreadParentId,
-} from '../../../views/streams/timelinesModel'
+} from '../../../views/models/timelineTypes'
+import { ETH_ADDRESS } from '@towns-protocol/web3'
+import { BlockchainTransaction_Tip, PlainMessage } from '@towns-protocol/proto'
 import { hexToBytes } from 'ethereum-cryptography/utils'
-import { getFallbackContent } from '../../../views/streams/timelineEvents'
+import { getFallbackContent } from '../../../views/models/timelineEvent'
 import { bin_toHexString } from '@towns-protocol/dlog'
 import { randomBytes } from '../../../utils'
