+++ packages/sdk/src/views/models/timelineTypes.ts
@@ -21,8 +21,8 @@ import {
     ChannelMessage_PostSchema,
     MembershipReason,
 } from '@towns-protocol/proto'
-import type { DecryptionSessionError } from '../../../decryptionExtensions'
-import { isDefined, logNever } from '../../../check'
+import type { DecryptionSessionError } from '../../decryptionExtensions'
+import { isDefined, logNever } from '../../check'
 import { create } from '@bufbuild/protobuf'
 
 export enum EventStatus {
@@ -612,4 +612,42 @@ export function transformAttachments(attachments?: Attachment[]): ChannelMessage
             }
         })
         .filter(isDefined)
+}
+
+export function getEditsId(content: TimelineEvent_OneOf | undefined): string | undefined {
+    return content?.kind === RiverTimelineEvent.ChannelMessage ? content.editsEventId : undefined
+}
+
+export function getRedactsId(content: TimelineEvent_OneOf | undefined): string | undefined {
+    return content?.kind === RiverTimelineEvent.RedactionActionEvent
+        ? content.refEventId
+        : undefined
+}
+
+export function getThreadParentId(content: TimelineEvent_OneOf | undefined): string | undefined {
+    return content?.kind === RiverTimelineEvent.ChannelMessage
+        ? content.threadId
+        : content?.kind === RiverTimelineEvent.TokenTransfer
+          ? content.threadParentId
+          : undefined
+}
+
+export function getReplyParentId(content: TimelineEvent_OneOf | undefined): string | undefined {
+    return content?.kind === RiverTimelineEvent.ChannelMessage ? content.replyId : undefined
+}
+
+export function getReactionParentId(content: TimelineEvent_OneOf | undefined): string | undefined {
+    return content?.kind === RiverTimelineEvent.Reaction ? content.targetEventId : undefined
+}
+
+export function getIsMentioned(content: TimelineEvent_OneOf | undefined, userId: string): boolean {
+    //TODO: comparison below should be changed as soon as this HNT-1576 will be resolved
+    return content?.kind === RiverTimelineEvent.ChannelMessage
+        ? content.mentions.findIndex(
+              (x) =>
+                  (x.userId ?? '')
+                      .toLowerCase()
+                      .localeCompare(userId.toLowerCase(), undefined, { sensitivity: 'base' }) == 0,
+          ) >= 0
+        : false
 }
