+++ packages/sdk/src/views/streams/spaceMentionsTransform.ts
@@ -1,6 +1,6 @@
 import { isEqual } from 'lodash-es'
 import { isChannelStreamId, spaceIdFromChannelId } from '../../id'
-import { MentionResult } from '../../sync-agent/timeline/models/timeline-types'
+import { MentionResult } from '../models/timelineTypes'
 import { TimelinesViewModel } from './timelinesModel'
 import { UnreadMarkersModel } from './unreadMarkersTransform'
