+++ packages/sdk/src/views/streams/timelines.ts
@@ -1,5 +1,5 @@
 import { Observable } from '../../observable/observable'
-import { TimelineEvent, RiverTimelineEvent } from '../../sync-agent/timeline/models/timeline-types'
+import { TimelineEvent, RiverTimelineEvent } from '../models/timelineTypes'
 import { LocalTimelineEvent, StreamTimelineEvent } from '../../types'
 import {
     makeTimelinesViewInterface,
@@ -7,7 +7,7 @@ import {
     TimelinesViewModel,
 } from './timelinesModel'
 import { StreamChange } from '../../streamEvents'
-import { toDecryptedContentErrorEvent, toDecryptedEvent, toEvent } from './timelineEvents'
+import { toDecryptedContentErrorEvent, toDecryptedEvent, toEvent } from '../models/timelineEvent'
 import { DecryptedContent } from '../../encryptedContentTypes'
 import { DecryptionSessionError } from '../../decryptionExtensions'
 import { isEqual } from 'lodash-es'
