+++ packages/sdk/src/views/streamsView.ts
@@ -19,7 +19,7 @@ import { ChannelStreamsView } from './streams/channelStreams'
 import { DmStreamsView } from './streams/dmStreams'
 import { GdmStreamsView } from './streams/gdmStreams'
 import { membershipsTransform } from './transforms/membershipsTransform'
-import { Membership } from '../sync-agent/timeline/models/timeline-types'
+import { Membership } from './models/timelineTypes'
 import { spaceIdsTransform } from './transforms/spaceIdsTransform'
 import { DmAndGdmModel, dmsAndGdmsTransform } from './transforms/dmsAndGdmsTransform'
 import { StreamMemberIdsView } from './streams/streamMemberIds'
